src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.5.2
azure-cosmos==4.5.1
python-jose[cryptography]==3.3.0
azure-storage-blob==12.19.0
pdfminer.six==20221105
Pillow==10.1.0
python-logging==0.4.9.6
